# backend/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SalesMate Backend API
  version: 0.2.0
  description: SalesMate Gmail understanding contract, company workspace and explicit
    rule placeholder.
paths:
...
  /api/v1/agent/chat/answers/:
    post:
      operationId: agent_chat_answer
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - agentCredential: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/agent/chat/context/:
    post:
      operationId: agent_chat_context
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - agentCredential: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/agent/chat/requests/claim/:
    post:
      operationId: agent_chat_claim
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - agentCredential: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/v1/sales/chat/messages/:
    post:
      operationId: chat_submit
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - laboratoryIdentity: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '201':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/sales/chat/requests/:
    get:
      operationId: chat_requests_list
      tags:
      - v1
      security:
      - laboratoryIdentity: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/sales/chat/requests/{request_id}/:
    get:
      operationId: chat_request_read
      parameters:
      - in: path
        name: request_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      security:
      - laboratoryIdentity: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/v1/sales/chat/requests/{request_id}/retry/:
    post:
      operationId: chat_retry
      parameters:
      - in: path
        name: request_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: {}
          application/x-www-form-urlencoded:
            schema:
              type: object
              additionalProperties: {}
          multipart/form-data:
            schema:
              type: object
              additionalProperties: {}
      security:
      - laboratoryIdentity: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
components:
...
  securitySchemes:
    agentCredential:
      type: apiKey
      in: header
      name: Authorization
      description: Agent <service-token>; never a Gmail token.
    businessToolCredential:
      type: apiKey
      in: header
      name: Authorization
      description: Tool <user-scoped-token>
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    laboratoryIdentity:
      type: apiKey
      in: header
      name: X-Lab-User
      description: Optional laboratory owner username, available only when LAB_OPEN_ACCESS=true; no token required. Omitting it uses the default laboratory identity. Unavailable in normal mode.